# src/services/report/routes/v1/public/doc/openapi.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "ezReeport API",
    "version": "1.0",
    "contact": {
      "name": "ezTeam",
      "url": "https://github.com/ezpaarse-project",
      "email": "ezpaarse@couperin.org"
    },
    "license": {
      "name": "CeCILL",
      "url": "https://github.com/ezpaarse-project/ezreeport/blob/master/LICENSE.txt"
    },
    "description": "Reporting service"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Direct"
    },
    {
      "url": "/report/api/v1",
      "description": "ezMESURE"
    }
  ],
  "paths": {
...
    "/me/namespaces": {
      "get": {
        "summary": "Get accessible namespaces",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": [
                        {
                          "id": "681d7a77-9551-4466-a48f-4a5e7f57d7d4",
                          "name": "Test3 Dev",
                          "logoId": null,
                          "createdAt": "2023-04-11T13:49:06.535Z",
                          "updatedAt": "2023-04-11T13:49:06.535Z"
                        },
                        {
                          "id": "a2c42f69-c2a0-436e-9099-ecd0fc4f73e6",
                          "name": "Test2 Dev",
                          "logoId": null,
                          "createdAt": "2023-04-11T13:48:39.383Z",
                          "updatedAt": "2023-04-11T13:48:39.383Z"
                        }
                      ]
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Namespace"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "_count": {
                                "type": "object",
                                "properties": {
                                  "tasks": {
                                    "type": "number"
                                  }
                                }
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "get-me-namespaces",
        "description": "List all accessible namespaces by user.\n\nNeeds the `namespaces[namespaceId].auth-get-namespaces` (see `/me/permissions` for more info), but it should be always granted."
      }
    },
...
    "/admin/namespaces": {
      "get": {
        "summary": "Get all namespaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Namespace"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "_count": {
                                "type": "object",
                                "properties": {
                                  "tasks": {
                                    "type": "number"
                                  },
                                  "memberships": {
                                    "type": "number"
                                  }
                                }
                              }
                            }
                          }
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "description": "Count of items available"
                        },
                        "count": {
                          "type": "number",
                          "description": "Count of items in response"
                        },
                        "size": {
                          "type": "number",
                          "description": "Count of items available"
                        },
                        "lastId": {
                          "type": "string",
                          "description": "Id of last item in response",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "get-admin-namespaces",
        "security": [
          {
            "API Key": []
          }
        ],
        "description": "Get all available namespaces.\n\nOnly usable with the API Key provided at server initialization.",
        "tags": [
          "admin - namespaces"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/previous"
          },
          {
            "$ref": "#/components/parameters/count"
          }
        ]
      },
      "post": {
        "summary": "Add new namespace",
        "operationId": "post-admin-namespaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/NamespaceFull"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin - namespaces"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NamespaceBody"
                  },
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "description": "Create a new namespace.\n\nOnly usable with the API Key provided at server initialization.",
        "security": [
          {
            "API Key": []
          }
        ],
        "deprecated": true
      },
      "put": {
        "summary": "Replace all namespaces",
        "operationId": "put-admin-namespaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "object",
                      "properties": {
                        "namespaces": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": [
                                  "created",
                                  "updated",
                                  "deleted"
                                ]
                              },
                              "data": {
                                "$ref": "#/components/schemas/Namespace"
                              }
                            }
                          }
                        },
                        "members": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": [
                                  "created",
                                  "updated",
                                  "deleted"
                                ]
                              },
                              "data": {
                                "allOf": [
                                  {
                                    "$ref": "#/components/schemas/Membership"
                                  },
                                  {
                                    "type": "object",
                                    "properties": {
                                      "username": {
                                        "type": "string"
                                      },
                                      "namespaceId": {
                                        "type": "string"
                                      }
                                    }
                                  }
                                ]
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "API Key": []
          }
        ],
        "description": "Create, update or delete many namespaces.\n\nIf you provide no `members`, it will keep current ones. If you provide empty `members`, it will remove all current memberships.\n\nWhen creating a membership, **user and namespace must exist**, so you may want to create users first (without `memberships`)\n\nOnly usable with the API Key provided at server initialization.",
        "tags": [
          "admin - namespaces"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/NamespaceBody"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "members": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/Membership"
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "username": {
                                    "type": "string"
                                  }
                                }
                              }
                            ]
                          }
                        }
                      },
                      "required": [
                        "id"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/namespaces/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get specific namespace",
        "tags": [
          "admin - namespaces"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/NamespaceFull"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-namespaces-id",
        "security": [
          {
            "API Key": []
          }
        ],
        "description": "Get specific namespace.\n\nOnly usable with the API Key provided at server initialization.",
        "parameters": []
      },
      "put": {
        "summary": "Update or create specific namespace",
        "operationId": "put-namespaces-id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/NamespaceFull"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {}
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/NamespaceFull"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "API Key": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceBody"
              }
            }
          }
        },
        "tags": [
          "admin - namespaces"
        ],
        "description": "Update specific namespace, if not found create it.\n\nOnly usable with the API Key provided at server initialization."
      },
      "delete": {
        "summary": "Delete specific namespace",
        "operationId": "delete-namespaces-id",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "API Key": []
          }
        ],
        "tags": [
          "admin - namespaces"
        ],
        "description": "Delete specific namespace.\n\nOnly usable with the API Key provided at server initialization."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Namespace": {
        "title": "Namespace",
        "x-stoplight": {
          "id": "69bji1pmw43zv"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "logoId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "User Token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used by user to interact with service"
      },
      "API Key": {
        "name": "X-API-Key",
        "type": "apiKey",
        "in": "header",
        "description": "Used by linked application to manage service"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "JWT malformed: invalid signature"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'GET /users/' requires auth"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'...' requires auth"
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User 'user' doesn't have the rights to acces to 'GET /users/'"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User '...' doesn't have the rights to access to '...'"
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "Unexpected token v in JSON at position 0"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "previous": {
        "name": "previous",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "0000a00a-aaa0-0000-a0aa-000a0a0aa000"
        },
        "description": "Previous id of the last request"
      },
      "count": {
        "name": "count",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "pattern": "^[0-9]+",
          "default": "15"
        },
        "description": "Count of object to return"
      },
...
    }
  },
...
}